Add caching mechanism for Hamiltonian computations in center manifold functions
- Implemented caching in center_manifold_rn and center_manifold_cn to avoid redundant calculations, improving performance.
- Enhanced logging to provide insights into cache usage and computation processes.
- Introduced methods for managing Hamiltonian and generating functions caches in the LibrationPoint class, facilitating efficient data retrieval and storage.
- This update optimizes the overall efficiency of Hamiltonian evaluations and transformations in the center manifold context.
