+++ src/algorithms/center/transforms.py
@@ -4,7 +4,6 @@ import numpy as np
 from numba import njit
 from numba.typed import List
 
-from algorithms.center.poincare.map import solve_p3
 from algorithms.center.polynomial.base import (_create_encode_dict_from_clmo,
                                                decode_multiindex)
 from algorithms.center.polynomial.operations import (polynomial_add_inplace,
@@ -271,305 +270,3 @@ def rn2phys(point, poly_rn: List[np.ndarray], max_deg: int, psi, clmo) -> List[n
     _, Cinv = point.normal_form_transform()
     encode_dict_list = _create_encode_dict_from_clmo(clmo)
     return substitute_linear(poly_rn, Cinv, max_deg, psi, clmo, encode_dict_list)
-
-
-def _cn2rn_coordinates(
-    cn_coords: np.ndarray,
-    max_degree: int,
-    psi: np.ndarray,
-    clmo: np.ndarray
-) -> np.ndarray:
-    """Convert complex normal coordinates to real normal coordinates."""
-    # Create a polynomial representation of the CN coordinates
-    cn_polys = polynomial_zero_list(max_degree, psi)
-    encode_dict_list = _create_encode_dict_from_clmo(clmo)
-    
-    # Set degree-1 terms to the coordinate values
-    if len(cn_polys) > 1:
-        for i in range(6):
-            if abs(cn_coords[i]) > 1e-15:
-                # Find position of x_i monomial in degree-1 polynomial
-                k = np.zeros(6, dtype=np.int64)
-                k[i] = 1
-                from algorithms.center.polynomial.base import encode_multiindex
-                pos = encode_multiindex(k, 1, encode_dict_list)
-                if 0 <= pos < cn_polys[1].shape[0]:
-                    cn_polys[1][pos] = cn_coords[i]
-    
-    # Use existing transformation function
-    rn_polys = cn2rn(cn_polys, max_degree, psi, clmo)
-    
-    # Extract coordinate values (degree-1 terms)
-    rn_coords = np.zeros(6, dtype=np.float64)
-    if len(rn_polys) > 1:
-        for i in range(6):
-            k = np.zeros(6, dtype=np.int64)
-            k[i] = 1
-            pos = encode_multiindex(k, 1, encode_dict_list)
-            if 0 <= pos < rn_polys[1].shape[0]:
-                rn_coords[i] = rn_polys[1][pos].real
-                
-    return rn_coords
-
-
-def _rn2phys_coordinates(
-    rn_coords: np.ndarray,
-    point,
-    max_degree: int,
-    psi: np.ndarray,
-    clmo: np.ndarray
-) -> np.ndarray:
-    """Convert real normal coordinates to physical coordinates."""
-    # Create polynomial representation
-    rn_polys = polynomial_zero_list(max_degree, psi)
-    encode_dict_list = _create_encode_dict_from_clmo(clmo)
-    
-    # Set degree-1 terms
-    if len(rn_polys) > 1:
-        for i in range(6):
-            if abs(rn_coords[i]) > 1e-15:
-                k = np.zeros(6, dtype=np.int64)
-                k[i] = 1
-                from algorithms.center.polynomial.base import encode_multiindex
-                pos = encode_multiindex(k, 1, encode_dict_list)
-                if 0 <= pos < rn_polys[1].shape[0]:
-                    rn_polys[1][pos] = rn_coords[i]
-    
-    phys_polys = rn2phys(point, rn_polys, max_degree, psi, clmo)
-    
-    # Extract physical coordinates
-    phys_coords = np.zeros(6, dtype=np.float64)
-    if len(phys_polys) > 1:
-        for i in range(6):
-            k = np.zeros(6, dtype=np.int64)
-            k[i] = 1
-            pos = encode_multiindex(k, 1, encode_dict_list)
-            if 0 <= pos < phys_polys[1].shape[0]:
-                phys_coords[i] = phys_polys[1][pos].real
-                
-    return phys_coords
-
-
-def _cm2phys_coordinates(
-    point,
-    cm_coords: np.ndarray,
-    poly_G_total: List[np.ndarray],
-    psi: np.ndarray,
-    clmo: np.ndarray,
-    max_degree: int,
-    energy: float = 0.0,
-    tol: float = 1e-15
-) -> np.ndarray:
-    """
-    Transform coordinates from center manifold back to physical (rotating) frame.
-    
-    This reverses the transformation pipeline used in center_manifold_cn:
-    Physical → RN → CN → Normalized CN → Center Manifold
-    
-    We apply the inverse: Center Manifold → CN → RN → Physical
-    
-    Parameters
-    ----------
-    point : object
-        Point object containing equilibrium point information and transformations
-    cm_coords : numpy.ndarray
-        Center manifold coordinates, either [q2, p2] for Poincaré section 
-        or [q2, p2, q3, p3] for full 4D coordinates
-    poly_G_total : List[numpy.ndarray]
-        Generating functions from lie_transform normalization
-    psi, clmo : arrays
-        Polynomial index tables
-    max_degree : int
-        Maximum degree used in normalization
-    energy : float, optional
-        Energy level for completing missing coordinates
-    tol : float, optional
-        Tolerance for cleaning small coefficients
-        
-    Returns
-    -------
-    numpy.ndarray
-        Physical coordinates [X, Y, Z, PX, PY, PZ]
-    """
-    
-    # Step 1: Complete center manifold coordinates if needed
-    full_cm_coords = _complete_cm_coordinates(
-        cm_coords, energy, point, psi, clmo, max_degree
-    )
-    
-    # Step 2: Apply inverse Lie transforms to get complex normal form
-    cn_coords = _apply_inverse_lie_transforms(
-        full_cm_coords, poly_G_total, psi, clmo, max_degree, tol
-    )
-    
-    # Step 3: Transform CN → RN → Physical using existing infrastructure
-    rn_coords = _cn2rn_coordinates(cn_coords, max_degree, psi, clmo)
-    physical_coords = _rn2phys_coordinates(rn_coords, point, max_degree, psi, clmo)
-    
-    return physical_coords
-
-
-
-@njit(fastmath=FASTMATH, cache=True)
-def _evaluate_reduced_monomial(
-    k: np.ndarray,
-    coords: np.ndarray, 
-    var_idx: int,
-    exp_change: int
-) -> np.complex128:
-    """Evaluate monomial x^k at coords with exponent of var_idx changed by exp_change."""
-    result = 1.0 + 0.0j
-    
-    for i in range(6):
-        exp = k[i]
-        if i == var_idx:
-            exp += exp_change
-            
-        if exp > 0:
-            if abs(coords[i]) > 1e-15:
-                result *= coords[i] ** exp
-            else:
-                return 0.0 + 0.0j  # Zero coordinate with positive exponent
-        # exp == 0: multiply by 1 (no-op)
-                
-    return result
-
-
-@njit(fastmath=FASTMATH, cache=True)
-def _compute_pb_coord(
-    G_coeffs: np.ndarray,
-    degree: int,
-    coord_idx: int,
-    coords: np.ndarray,
-    clmo: np.ndarray,
-    encode_dict_list: List
-) -> np.complex128:
-    """
-    Compute {G, x_coord_idx} evaluated at the given coordinates.
-    
-    For a monomial G_term = coeff * q^a * p^b:
-    {G_term, q_i} = -coeff * b_i * q^a * p^(b with p_i reduced by 1)  
-    {G_term, p_i} = +coeff * a_i * q^(a with q_i reduced by 1) * p^b
-    """
-
-    result = 0.0 + 0.0j
-    
-    for pos in range(G_coeffs.shape[0]):
-        coeff = G_coeffs[pos]
-        if abs(coeff) < 1e-15:
-            continue
-            
-        # Decode the multi-index for this monomial
-        k = decode_multiindex(pos, degree, clmo)
-        
-        if coord_idx < 3:  # Position coordinate q_i
-            # {G_term, q_i} = -k[i+3] * G_term with p_i exponent reduced by 1
-            p_idx = coord_idx + 3
-            if k[p_idx] > 0:
-                # Evaluate monomial with p_i exponent reduced by 1
-                monomial_val = _evaluate_reduced_monomial(k, coords, p_idx, -1)
-                result -= coeff * k[p_idx] * monomial_val
-        else:  # Momentum coordinate p_i  
-            # {G_term, p_i} = +k[i] * G_term with q_i exponent reduced by 1
-            q_idx = coord_idx - 3
-            if k[q_idx] > 0:
-                # Evaluate monomial with q_i exponent reduced by 1
-                monomial_val = _evaluate_reduced_monomial(k, coords, q_idx, -1)
-                result += coeff * k[q_idx] * monomial_val
-                
-    return result
-
-
-@njit(fastmath=FASTMATH, cache=True)
-def _apply_single_inverse_generator(
-    coords: np.ndarray,
-    G_n: np.ndarray,
-    degree: int,
-    psi: np.ndarray,
-    clmo: np.ndarray,
-    encode_dict_list: List
-) -> np.ndarray:
-    """
-    Apply inverse of a single generating function using first-order approximation.
-    
-    For the inverse transformation, we use:
-    x_new = x_old - {G, x_old} + O(G^2)
-    
-    This is the first-order approximation of exp(-L_G).
-    """
-    new_coords = coords.copy()
-    
-    # Compute Poisson brackets {G, x_i} for each coordinate
-    for i in range(6):
-        pb_term = _compute_pb_coord(
-            G_n, degree, i, coords, clmo, encode_dict_list
-        )
-        new_coords[i] -= pb_term  # Negative sign for inverse transform
-    
-    return new_coords
-
-
-@njit(fastmath=FASTMATH, cache=True)
-def _apply_inverse_lie_transforms(
-    cm_coords: np.ndarray,
-    poly_G_total: List[np.ndarray],
-    psi: np.ndarray,
-    clmo: np.ndarray,
-    max_degree: int,
-    tol: float
-) -> np.ndarray:
-    """Apply inverse Lie transforms to go from center manifold to complex normal coordinates."""
-    encode_dict_list = _create_encode_dict_from_clmo(clmo)
-    
-    # Start with 6D coordinates: [q1=0, q2, q3, p1=0, p2, p3]
-    coords = np.zeros(6, dtype=np.complex128)
-    coords[1] = cm_coords[0]  # q2
-    coords[2] = cm_coords[2]  # q3
-    coords[4] = cm_coords[1]  # p2  
-    coords[5] = cm_coords[3]  # p3
-    # q1=0, p1=0 remain zero (center manifold constraint)
-    
-    # Apply inverse generating functions in reverse order
-    # The generating functions were applied in order G3, G4, ..., GN
-    # So we apply them in reverse: -GN, -G(N-1), ..., -G3
-    for degree in range(max_degree, 2, -1):
-        if degree < len(poly_G_total) and np.any(poly_G_total[degree]):
-            coords = _apply_single_inverse_generator(
-                coords, poly_G_total[degree], degree, psi, clmo, encode_dict_list
-            )
-    
-    return coords
-
-
-def _complete_cm_coordinates(
-    poly_cm: List[np.ndarray],
-    cm_coords: np.ndarray,
-    energy: float,
-    point,
-    psi: np.ndarray,
-    clmo: np.ndarray,
-    max_degree: int
-) -> np.ndarray:
-    """Complete center manifold coordinates using energy constraint."""
-    if len(cm_coords) == 4:
-        return cm_coords.astype(np.complex128)
-    elif len(cm_coords) == 2:
-        # Poincaré section case: solve for p3 using existing infrastructure
-        q2, p2 = cm_coords
-        q3 = 0.0
-        
-        # Use existing solve_p3 function from map.py
-        p3 = solve_p3(
-            q2=float(q2), 
-            p2=float(p2), 
-            h0=energy, 
-            H_blocks=poly_cm, 
-            clmo=clmo
-        )
-        
-        if p3 is None:
-            p3 = 0.0
-            
-        return np.array([q2, p2, q3, p3], dtype=np.complex128)
-    else:
-        raise
